back/requirements.txt
django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-cors-headers==4.4.0
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
drf-spectacular==0.28.0
